text_preprocessing/english/requirements.txt
click==8.1.7
joblib==1.3.2
nltk==3.8.1
pip==23.2.1
pyspellchecker==0.8.1
regex==2023.12.25
setuptools==65.5.0
tqdm==4.66.2
